Calculating salt balance from ion chromatography data (now deprecated for updated version)

Performs a salt balance on ion chromatography data and returns outputs for ECOS Runsalt software.

This function is now deprecated and is being replaced with a new function.

Usage

fun_salt_calcs_old(
  salt_input,
  sample_name,
  dry_g,
  water_ml,
  chloride_ppm,
  nitrate_ppm,
  sulfate_ppm,
  sodium_ppm,
  potassium_ppm,
  calcium_ppm,
  magnesium_ppm
)

Arguments

salt_input

Dataframe containing ion chromatography data

sample_name

Sample name

dry_g

Dry mass of sample, g

water_ml

Water added for ion chromatography, ml

chloride_ppm

Chloride concentration, ppm

nitrate_ppm

Nitrate concentration, ppm

sulfate_ppm

Sulfate concentration, ppm

sodium_ppm

Sodium concentration, ppm

potassium_ppm

Potassium concentration, ppm

calcium_ppm

Calcium concentration, ppm

magnesium_ppm

Magnesium concentration, ppm

Value

Dataframe of balanced ions for ECOS Runsalt software

Examples

salt_test |> fun_salt_calcs_old()
#> # A tibble: 1 × 148
#>   sample_name dry_g water_ml chloride_ppm nitrate_ppm sulfate_ppm sodium_ppm
#>   <chr>       <dbl>    <dbl>        <dbl>       <dbl>       <dbl>      <dbl>
#> 1 test         1.13      100         66.8        333.        87.2       21.5
#> # ℹ 141 more variables: potassium_ppm <dbl>, calcium_ppm <dbl>,
#> #   magnesium_ppm <dbl>, Chloride_wt <dbl>, Nitrate_wt <dbl>, Sulfate_wt <dbl>,
#> #   Sodium_wt <dbl>, Potassium_wt <dbl>, Calcium_wt <dbl>, Magnesium_wt <dbl>,
#> #   total_salt_wt <dbl>, chloride_mEq <dbl>, nitrate_mEq <dbl>,
#> #   sulfate_mEq <dbl>, sodium_mEq <dbl>, potassium_mEq <dbl>,
#> #   calcium_mEq <dbl>, magnesium_mEq <dbl>, total_anions_mEq <dbl>,
#> #   total_cations_mEq <dbl>, xs_cations_mEq <dbl>, ave_ions_mEq <dbl>, …